[#28] Explicit type of named functions
For readability and compilation optimization,
we want to explicit the return type of named function.

Currently, we always check the return type of the function body.
But when the return type is not the one expected, the debug is more complex.
Also, the compilation is slower and could be difficult to improve.

This commit do the following:

* The return type is now mandatory
* The named function does not use a function anymore (model changed)
* Change all calls to named function (read or write) for new model
* The return type of a named function is the one defined, not the body
* Correct sample file
* Adapt tests