[#28] Simplify the syntax of named function
For readability and writing purpose, simplify the syntax of named functions.

Currently, we need to add a `=` sign and a `=>` sign to define a function.
It is not obvious and was only used to keep similarity with lambda.
Now, the syntax is simplified by removing the `=` sign.

Also, because when the body contains only one expression,
forcing the usage of braces can be boilerplate.

This commit do the following:

* Remove the `=` sign in antlr file and sample file
* Remove the mandatory use of braces when there is only one expression
* Adapt antlr file, parser and tests to accept both versions