[#23] Create enhanced AST for public API
Because the public AST should be more specific and hide internal needs,
we create an enhanced AST adding information on the source AST.

We create `Project.generatePublicAST` instead of updating `Project.process`.
We will update this process afterward.
The objective is to be able to test the process creating public AST.

This commit do the following:

* Create `typed` and `structure` AST
* Add two other steps in Project to include information into new AST
* `DefinedMethodDefinition` is never used currently, delete it

Remark: Currently, we check twice types in expression (validation and typing).
A following commit will correct this behavior.