[#23] Refactor the whole verification part
With the issue #23, we include the typing into the AST.
At the moment, the steps were validation then typing.
But by design the typing verifies one more time all expressions and typing.
To avoid code duplicate and simplify the process, we put the validation after the typing.

The AST was totally changed so the validation must also change.
Doing so, we discover some issues:

* Lot of validations was not done or was invalid
* The structure of validation could be improved (so it was)
* The tests were not quite testable and reusable because they all needed a context not used
* The tests did not really test anything

The most important lesson learned from this refactoring is that PBT (Property-Based Testing)
is not the right strategy for this project.
Also we should avoid to test function and adopt a more end-to-end testing.

The consequence is that lot of tests were deleted because the whole structure changed.

**The strategy for tests from now will be to consider the whole compiler as a unit
and tests from source files to structure result.**

This commit do the following:

* Put the validation after structuring the AST
* Restructure the validation :
  * Splitting it into several traits
  * Consider the library as a property across the class instead of function parameter
  * Use the structure AST and not the pure AST
* Remove validation tests (more specific tests will be needed)
* Correct some validations and tests
* Remove unused or deprecated functions and types