[#23] Create nominal tests
Because we removed lot of deprecated tests, we need to add some others.
We start with nominal tests with the strict minimum elements.
We check for all top-level elements, without any complexity.

The implementation of these tests shows an existing bug for elements in root package.
The project gave two version of root-package elements.

Also, the `ValidationMatcher` did not give an explicit message when there was an error.

There is a need for more complex tests that should be added later.

This commit do the following:

* Correct the project structure to not create a sub-package for root package
* Extract common code for end-to-end tests to its trait `EndToEndSpec`
* Create nominal tests with the strict minimum
* Correct the `ValidationMatchers` to show the error in case of failure