[#23] Move public AST types to `definiti.core.ast`
Now the work in mainly done, we move the public API to `definiti.core.ast`.
The objective is to explicit private and public API.
The private API stays in `ast.pure` and `ast.typed`.

To avoid name collision which happened during refactoring
and was experienced during the development of this issue,
we also rename the private API with prefixes.

The impact is a huge correction of imports.

To enforce the public/private API, we correct visibility by adding
`private[core]` visibility to private API.

This commit do the following:

* Move the public API to `definiti.core.ast`
* Correct all the imports
* Avoid name collision
* Correct visibility
* Change `Package` to `Namespace` because of a bug with `java.lang.Package`