[#30] Correct the range and position of elements
Since the beginning, there was an error about position of errors.

The line count started from 1 but columns from 0.
We correct the logic so both of them start from 1 (as in editors).

Also, Antlr give as `stop` the last token, not the last position.
Because all tokens are on one line in this project,
we consider the last position being its position plus its size.

This commit do the following:

* The first index of column is now 1
* The end position is corrected
* Remove `TestConstants` and move methods into specific models
* Correct tests to reflect the new positions