[#30] Update AST model to show the Location instead of the Range
At the beginning, we had only one file, so showing the range was enough.
With the package and multiple file structure,
the range is no enough to show where an element came from.

The change from `Range` to `Location` was never made, so we do it now.

With this update, an error was detected.
In `ExpressionTyping.addTypeIntoAttributeCall`,
we copied the attributeDefinition `Range` instead of `attributeCall`.
We correct it at the same time.

This commit do the following:

* Change `range: Range` to `location: Location` in AST
* Propagate the change to the whole project
* Correct the copy of `attributeDefinition.range` to `attributeCall.location`
* Update the plugin model to use `Location` instead of `Range`
* Update the project structure to indicate the current file used
* Update tests with the new model