Correct generic types in named function
When in a named function there was a generic, the system was unable to process it.
It ended into an invalid result even though the code was valid.

We set an abstract type without any element when we encounter a generic.
The reason is because there is no covariance in the project.
In the context, we cannot determine the type at compile time (as in other languages).

This commit do the following:

* Add test of the failing case
* Add another test to check a real invalid scenario
* Correct validated matcher (in tested `Valid` instead of `ValidValue`)
* When the named function has a generic and is the one used, set it as abstract
* Before testing expression from a named function, use the named function context