Correct deep namespaces
When the number of namespaces was more than one, the system ignored them.
The reason was the detection of children of a package.
The method got the last part of parent package and tried to get its children with the full path.

Now, the full path is given and used to detect children packages.

This commit do the following:

* Add test of the failing case (sub-package)
* Add another test to better check recursion (sub-sub-package)
* Use the full path of package name, not only its last part