The `Project` and `Configuration` classes are now public
When implementing tests in `definiti/scala-model`, we encounter an issue:
It is impossible to access the `Project` class from outside.

This class is used to generate the public AST intended for test cases.
So we set this class public.

`Project` class needs `Configuration`, so we also set it public.

This commit does the following:

* Set `Project` visibility to `public`
* Set the version number as `0.2.0-snapshot` instead of keeping it `0.1.0` in version control