[#43] Accept executables instead of scala plugins only
Because Definiti should be used to generate codes in several languages,
having a strong dependency on one language is not quite logic.

Now we can accept binary files to transform or validate AST or generate files.

The interface is nearly the same as scala plugin,
but as command line arguments instead of interface methods.

A try was done with JNA without success because of polymorphism.
We fallback to a version as arguments in json.

This commit does the following:

* Create JSON serialization for AST (pure and final) and library
* ContextPlugin interface need to change to define how to serialize json
* Create a class to create instance of command-line plugins
* Dependencies are accepted as command line with prefix `>` so we change configuration

Remark: Only tested into linux embedded in windows.
We should confirm it works in docker container.