[#50] Create `Program` helper
To continue evolution and improve validation system,
we create a new type `Program`.
At the end of this evolution, other types will be replaced by this one.
Currently, we just create transformers between types so the commit stays simple.

The objective of `Program` is to separate definition and execution.
Also, execution will be dependant of `Configuration` so we can configure execution steps.

The definition of `Program` is compatible with for comprehension.
So we can simplify `Project` structure.

This commit does the following:

* Create type `Program` and its specification
* Create type `Alert` to have several alerts (blocking, informative…)
* Use these types for each step of `Project` (via helpers)
* Split `CoreParser` and `ProjectParser` and flatten results in `Project`
* Flatten steps in `Project` with for comprehension
* Correct tests to use new types