[#50] Add `Control` system
Currently AST validation is done by going through the AST and validate it at each step.
This scenario is hard to read and maintain.

We change the strategy by creating a new notion: `Control`.

Instead of going through the AST to validate everything,
we define a list of controls and apply them to the AST.
These controls control only one thing.
They are configurable (level of alerts, and fatal alerts)
so we can easily add or remove control inside a project.

In this commit, we add controls without changing the existing code.
This work will be done afterward.

This commit does the following:

* Create trait `Control` used by all controls
* Create a control for each control level to represent the system
* Update `ASTValidation` to take into account these control
* Update configuration to make controls configurable
* Update `Program` to take into account blocking alerts depending on configuration
* Update example to represent the new system