[#57] Remove intermediate AST pure
With the previous commit, we reduced the number of AST to two.
The only real difference between these two AST is the type.
This commit remove the first AST to simplify code and model.

The type will be defined as `Unset` for the time being and will be
set during typing, like before.

This strategy remove type security to check the type was set.
However, the hypothesis is tests should help us avoid unset types.

This commit does the following:

* Remove pure AST
* Parser use directly final AST
* returnType of expression is always the Abstract version
* During typing, copy AST elements instead of creating a full new one
* Simplify `Context` by removing unused methods
* The common package has no dependency to core package