[#60] Parameter type can be implicitly extracted from its name - lambda
To simplify reading and writing, the parameter type definition can be implicit.
It will be extracted from the parameter name by capitalizing the name.

This commit focus on lambda parameters.

Previous commit already did the work so we just add tests to ensure the process.

This commit does the following:

* Add two tests to control the new behavior