[#73] Use scala parsing for definiti file parsing
We remove antlr usage and use scala parsing instead.
We already did the core definition so now we do the project itself.

We do not change the rest of the application so the impact is only on the parsing part.

As a result, we also corrected a bug with extended contexts.

This commit does the following:

* Remove the antlr definition for project parsing (scala code and build definition)
* Remove the previous parser using antlr version
* Define the `Token`s, `Lexer` and `Parser` for project
* Change some logic in the parsing phase:
  * Import processing were done during parsing but now it is done afterward
  * Extended contexts are processed afterward too (because of imports)